Polishing fastify example
