+++ examples/fastify/errors.ts
@@ -1,0 +1,13 @@
+export { InvalidCVC, UknownProduct, MissingCardNumber };
+
+class InvalidCVC {
+  __brand!: 'InvalidCVC';
+}
+
+class UknownProduct {
+  __brand!: 'UknownProduct';
+}
+
+class MissingCardNumber {
+  __brand!: 'MissingCardNumber';
+}
